Fix apn overwrite issue for modem SSR case
When an apn is opened for edit, and modem SSR is performed,
apneditor is losing focus, due to which it is paused and resumed
multiple times. Apn fields are getting reset in onCreate() method
which is invoked when apneditor is resumed. Due to this, apn is
getting overwritten with empty values.

Fix is to read apn fields from apn database always.
Remove mFirstTime flag from ApnEditor class.

Change-Id: Ib71a49527e6ff12e9db6e3bbdf732f89313376dd
CR-Fixed: 2045973